+++ src/Components/ModalEndpoint.svelte
@@ -1,20 +1,36 @@
 <script>
     import EndpointInfo from '../Types/EndpointInfo';
-    import { saveEndpoint } from '../Services/EndpointsService';
-    import {UiConstants} from '../Types/Constants';
+    import { saveEndpoint, endpointExists } from '../Services/EndpointsService';
+    import { UiConstants } from '../Types/Constants';
+    import { showErrorToast, showSuccessToast } from '../Auxiliaries/UiHelpers'
     
-    
+    let https = '';
     let endpoint = new EndpointInfo();
 
     export function showModal() {
       endpoint = new EndpointInfo();
+      https = 'https://';
       window.$(UiConstants.ModalEndpoint_IdSelector).modal('show');
     }
 
 
     async function uiOnSaveClicked() {
-        //TODO validate form
+        //<Validation>
+        if (!endpoint.description) {
+          showErrorToast('Description is required');
+          return;
+        }
+        const exists = await endpointExists(endpoint.description);
+        if (exists) {
+          showErrorToast('Description exists');
+          return;
+        }
+        //</Validation>
+
+
+        endpoint.url = https + endpoint.url;
         await saveEndpoint(endpoint);
+        await showSuccessToast('endpoint created');
         //close modal
         window.$(UiConstants.ModalEndpoint_IdSelector).modal('hide');
     }
@@ -30,13 +46,23 @@
 
   <div class="content">
     <form class="ui form">
+      <!-- Row-1: Description-->
       <div class="field">
         <label for="description">Description</label>
         <input bind:value={endpoint.description} type="text" class="form-control" id="description" placeholder="Description">
       </div>
-      <div class="field">
-        <label for="url">Url</label>
-        <input bind:value={endpoint.url} type="text" class="form-control" id="url" placeholder="Url">
+      <!-- Row-2-->
+      <h4 class="ui dividing header">Url</h4>
+      <div class="fields">
+        <div class="two wide field">
+          <select class="ui fluid search dropdown" bind:value={https}>
+            <option value="https://">https://</option>
+            <option value="http://">http://</option>
+          </select>
+        </div>
+        <div class="fifteen wide field">
+          <input bind:value={endpoint.url} type="text" class="form-control" id="url" placeholder="Url">
+        </div>
       </div>
     </form>
   </div>