+++ ChangeLog.md
@@ -1,3 +1,12 @@
+## [0.0.4] - yyy-mm-dd
+### Added
+- [AddEndpoint]: verify description is unique
+- [AddEndpoint]: add http/https dropdown
+ 
+### Changed
+ 
+### Fixed
+
 ## [0.0.3] - 2023-02-17
 
 ### Added